Add logger manager and configuration
Create a LoggerManager that serves as a starting point for new loggers
in a custom application and manages a shared configuration for loggers
(including backend and logging level).

Additionally implement a hierarchial logging level that 'inherits' the
parent's configuration. This is useful e.g. when individual loggers
should use a different level than the global configuration.

Finally add some more documentation comments, especially in the Logger
interface.